fix: move setPosterHeaderTimeout to separate file to satisfy API file size guardrail
