+++ internal/api/core/poster_timeout.go
@@ -1,0 +1,25 @@
+package core
+
+import (
+	"context"
+	"net"
+	"net/http"
+	"time"
+)
+
+func setPosterHeaderTimeout(client *http.Client, idleTimeout time.Duration) {
+	if t, ok := client.Transport.(*http.Transport); ok && idleTimeout > 0 {
+		t.ResponseHeaderTimeout = idleTimeout
+		t.TLSHandshakeTimeout = idleTimeout
+		if t.DialContext != nil {
+			origDial := t.DialContext
+			t.DialContext = func(ctx context.Context, network, addr string) (net.Conn, error) {
+				dialCtx, cancel := context.WithTimeout(ctx, idleTimeout)
+				defer cancel()
+				return origDial(dialCtx, network, addr)
+			}
+		} else {
+			t.DialContext = (&net.Dialer{Timeout: idleTimeout}).DialContext
+		}
+	}
+}
